creating strain fitting mixin for plotting and calculation
